BUGFIX:  2481213  	 3.0.2 - Dada Mail broken without DBI module https://sourceforge.net/tracker2/?func=detail&aid=2481213&group_id=13002&atid=113002
The problem seemed to be that DBI is always attempted to be used, no matter if we want to use an SQL backend or not. 

This bug probably cropped up, since I started taking out the oft-repeating code that checks if we're using it - that's a good thing, getting rid of that code, but at least one instance of something similar has to be in place - and the DADA::App::DBIHandle module is a good place for that.